+++ backend/src/main/java/com/roc/app/competition/exception/CompetitionNotFoundException.java
@@ -1,7 +1,7 @@
 package com.roc.app.competition.exception;
 
 public class CompetitionNotFoundException extends RuntimeException {
-    public CompetitionNotFoundException(Long id) {
+    public CompetitionNotFoundException(Integer id) {
         super("Competition not found with id: " + id);
     }
 }
